pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "pyseval"
version = "0.1.0"
description = "A CLI application that validates power bi semantic models by comparing actual dax query results obtained from semantic models in the power bi service using the power bi api with their expected results."
authors = [
    {name = "Malte Brunst", email = "malte.brunst@m-bi.com"},
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = "~=3.12"
keywords = ["powerbi", "semantic model", "testing", "validation"]
dependencies = [
    "pandas ~=2.2",
    "jsonschema ~=4.21",
    "requests ~=2.31",
    "azure-identity ~=1.15",
    "click ~=8.1",
    "python-dotenv ~=1.0"
]

[project.scripts]
pyseval = "pyseval.cli:cli"

[project.urls]
Source = "https://github.com/m-bi-github/pyseval"

[tool.pytest.ini_options]
filterwarnings = [
    'ignore::DeprecationWarning:dateutil.tz:37',
]